seperate models and create common module
