+++ mail_reset/models/mail_reset_domain.py
@@ -1,30 +1,7 @@
 # -*- coding: utf-8 -*-
 
+from . import common
 from odoo import models, fields, api
-from odoo.exceptions import Warning
-
-import requests
-
-
-def _check_api_rights(api_url, api_token, namespace, verb, resource):
-    hed = {'Authorization': 'Bearer ' + api_token}
-    data = {"kind":"SelfSubjectAccessReview",
-            "apiVersion":"authorization.k8s.io/v1",
-            "metadata":{"creationTimestamp":None},
-            "spec":{"resourceAttributes":{
-                "namespace":namespace,
-                "verb": verb,
-                "resource":resource}},
-            "status":{"allowed":False}}
-        
-    url = f'{api_url}/apis/authorization.k8s.io/v1/selfsubjectaccessreviews'
-    response = requests.post(url, json=data, headers=hed, verify=False)
-    output = response.json()
-    if output['status'] == 'Failure':
-        raise Warning('You are NOT authorized!')
-    else:
-        return output['status']['allowed']
-
 
 class Mail_Reset_Domain(models.Model):
     _name = 'mail_reset.domain'
@@ -50,6 +27,6 @@ class Mail_Reset_Domain(models.Model):
              ('*','pods/exec'),
             ]
         for verb, resource in A:
-            if _check_api_rights(self.api_url, self.api_token, self.namespace, verb, resource) == False:
+            if common._check_api_rights(self.api_url, self.api_token, self.namespace, verb, resource) == False:
                 return self.write({'state': 'fail'})
         return self.write({'state': 'success'})
